scala-js-part2: addressed review comments
