+++ interpreter/js/src/main/scala/sigmastate/js/Isos.scala
@@ -4,7 +4,9 @@ import sigmastate.{Iso, SType}
 import sigmastate.Values.Constant
 import sigmastate.eval.Evaluation
 
+/** Definitions of isomorphisms. */
 object Isos {
+  /** Conversion between `Value` and `Constant[SType]`. */
   implicit val isoValueToConstant: Iso[Value, Constant[SType]] = new Iso[Value, Constant[SType]] {
     override def to(x: Value): Constant[SType] =
       Constant(x.runtimeData.asInstanceOf[SType#WrappedType], Evaluation.rtypeToSType(x.tpe.rtype))
