+++ project/plugins.sbt
@@ -10,6 +10,6 @@ addSbtPlugin("com.github.sbt" % "sbt-pgp" % "2.2.1")
 addSbtPlugin("com.dwijnand" % "sbt-dynver" % "4.1.1")
 
 addSbtPlugin("org.portable-scala" % "sbt-scalajs-crossproject" % "1.2.0")
-addSbtPlugin("org.scala-js"       % "sbt-scalajs"              % "1.11.0")
+addSbtPlugin("org.scala-js"       % "sbt-scalajs"              % "1.10.1")
 addSbtPlugin("ch.epfl.scala"      % "sbt-scalajs-bundler"      % "0.20.0")
 addSbtPlugin("org.scalablytyped.converter" % "sbt-converter" % "1.0.0-beta37")
